requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
billiard==4.2.0
boto3==1.16.63
botocore==1.19.63
CacheControl==0.13.1
cachetools==5.3.2
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.7
Deprecated==1.2.14
elasticsearch==8.13.0   # TODO: Evaluate the version
ethos>=24.6.127
ethosdev>=24.6.7
diskcache==5.6.3    # TODO: Evaluate the need
firebase-admin==6.3.0
FLAML==2.1.1    # TODO: Evaluate the need
frozenlist==1.4.0
google-api-core==2.15.0
google-api-python-client==2.110.0
google-auth==2.25.2
google-auth-httplib2==0.1.1
google-cloud-core==2.4.1
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.62.0
greenlet==3.0.2
grpcio==1.60.0
grpcio-health-checking==1.60.0
grpcio-status==1.60.0
grpcio-tools==1.60.0
httplib2==0.22.0
idna==3.6
importlib-metadata==6.11.0
jaeger-client==4.8.0
jmespath==0.10.0
kombu==5.3.4
msgpack==1.0.7
multidict==6.0.4
numpy==1.26.2
openai==1.6.0   # TODO: Evaluate the need
opencv-python==4.8.1.78
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-exporter-otlp-proto-http==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-grpc==0.42b0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
opentracing==2.4.0
packaging==23.2
pexpect==4.9.0      # TODO: Evaluate the version
pdf2image==1.16.3
phonenumbers==8.13.26
Pillow==10.1.0
prompt-toolkit==3.0.41
proto-plus==1.23.0
protobuf==4.25.1
psycopg2-binary==2.9.9
py3dns==4.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyautogen==0.2.23   # TODO: Evaluate the need
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pyfcm==1.5.4
PyJWT==2.8.0
pyparsing==3.1.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0    # TODO: Evaluate the need
python-http-client==3.3.7
redis==5.0.1
requests==2.31.0
rsa==4.9
s3transfer==0.3.7
scikit-learn==1.4.0     # TODO: Evaluate the version
scipy==1.12.0   # TODO: Evaluate the version
sendgrid==6.11.0
six==1.16.0
SQLAlchemy==2.0.23
starkbank-ecdsa==2.2.0
stripe==7.8.1
tiktoken==0.5.2     # TODO: Evaluate the need
threadloop==1.0.2
thrift==0.16.0
torch==2.2.2    # TODO: Evaluate the version
tornado==6.4
transformers==4.39.3     # TODO: Evaluate the version
twilio==8.10.3
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.18
validate-email==1.3
vine==5.1.0
wcwidth==0.2.12
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0